refactor(tool): Move microVM builder to utils.rs
Move a functionality for building a microVM from the given config file
to `utils.rs` as `build_microvm_from_config()`.

Resource preparation for building a microVM is not directly related to
the dump feature and can be considered re-usable and one of shared
utilities.
